Nsxv Enhancements
The header options for the nsx policy will support securitygroup securitygroup-id as the target options. These are optional. If these target options are given, all the rules in the section will have appliedTo as the given Security Group.
If the values are not given, the appliedTo will be default to DISTRIBUTED_FIREWALL