Automated g4 rollback of changelist 296920068.
*** Reason for rollback ***

Having an action map of next:goto_next in the gcp_hf.py lib is ok, so we don't need a new subtoken.  What we do not want is a map of something like accept:goto_next because it's changing the meaning of the ACL.

*** Original change description ***

Add a new subtoken: goto_next.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=297425358