Automated g4 rollback of changelist 331050394.
*** Reason for rollback ***

This was not the way to go. We don't want to drop the IPv6 protocols in cases when the only protocols in the list are IPv6 protocols because a term with no protocol (which translates to all protocols) will be created, causing an overexposure.  Another solution in the works.

*** Original change description ***

Add a way to ignore ICMPv4/6 the same way we can already ignore IPV4/6 adresses.

***

PiperOrigin-RevId: 331239278