Update gcp_hf to raise errors to when target resources or port list exceed 256 items.
Currently, when source-address or destination-address in HF terms exceeds 256 elements, the gcp_hf generator will split those rules into new rules that comply with this limit. With this change, if destination-ports and target-resources contain more than 256 entries, gcp_hf raises an error, so the user is expected to provide terms that do not exceed these limits. However, the future plan is to support splitting of these fields into different rules.

PiperOrigin-RevId: 340522557